fix: add exit code check for deploy.sh

+++ ci/deploy.sh
@@ -2,6 +2,12 @@
 
 python setup.py sdist upload -r http://scmesos06
 
+# release failed
+[ $? -gt 0 ] && exit 1
+
 pip uninstall -y sirius
 
 pip --trusted-host scmesos06 install -i http://scmesos06/simple sirius
+
+# install failed
+[ $? -gt 0 ] && exit 2
